Format newlines in surf.inc
One of the strings declared in surf.inc was poorly formatted.

## Description
Some of the strings declared in surf.inc contained newline characters in the middle of the .string line, opposing the format used by every string in the `data/text` directory, I just updated them to match the format, making it more readable.
# Conflicts:
#	data/text/surf.inc